removed jq vulnerbility. capped events time. removed extra files W
