+++ .gitignore
@@ -4,3 +4,4 @@ bower_components/
 *.log.csv
 client/packages.old
 client/libs/
+.DS_Store
