+++ package.json
@@ -10,15 +10,15 @@
   "author": "jmrnilsson",
   "license": "MIT",
   "dependencies": {
-    "express": "^4.13.3",
-    "serialport": "^4.0.1",
-    "socket.io": "^1.3.7",
     "bootstrap": "^3.3.6",
     "c3": "^0.4.11",
     "d3": "^3",
-    "jquery": "2.x",
+    "express": "^4.13.3",
+    "jquery": "^3.3.1",
     "knockout": "^3.4.0",
     "requirejs": "^2.2.0",
+    "serialport": "^4.0.1",
+    "socket.io": "^1.3.7",
     "socket.io-client": "^1.4.8",
     "text": "github:requirejs/text",
     "toastr": "^2.1.2"
