--- run.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-(./node_modules/.bin/eslint ./ --ext .js && node ./fdaemon.js)
